+++ pumpkin-lib/src/branching/value_selection/in_domain_max.rs
@@ -1,16 +1,16 @@
-use super::ValueSelector;
 use crate::basic_types::variables::IntVar;
 use crate::basic_types::Literal;
 use crate::branching::SelectionContext;
+use crate::branching::ValueSelector;
 
-/// [`ValueSelector`] which chooses to assign the provided variable to its lowest-bound.
+/// [`ValueSelector`] which chooses to assign the provided variable to its upper-bound.
 #[derive(Debug, Copy, Clone)]
-pub struct InDomainMin;
+pub struct InDomainMax;
 
-impl<Var: IntVar + Copy> ValueSelector<Var> for InDomainMin {
+impl<Var: IntVar + Copy> ValueSelector<Var> for InDomainMax {
     fn select_value(&mut self, context: &mut SelectionContext, decision_variable: Var) -> Literal {
         context.get_literal_for_predicate(
-            decision_variable.upper_bound_predicate(context.lower_bound(decision_variable)),
+            decision_variable.lower_bound_predicate(context.upper_bound(decision_variable)),
         )
     }
 }
@@ -19,7 +19,7 @@ impl<Var: IntVar + Copy> ValueSelector<Var> for InDomainMin {
 mod tests {
     use crate::basic_types::tests::TestRandom;
     use crate::basic_types::PredicateConstructor;
-    use crate::branching::InDomainMin;
+    use crate::branching::InDomainMax;
     use crate::branching::SelectionContext;
     use crate::branching::ValueSelector;
 
@@ -36,13 +36,13 @@ mod tests {
         );
         let domain_ids = context.get_domains().collect::<Vec<_>>();
 
-        let mut selector = InDomainMin;
+        let mut selector = InDomainMax;
 
         let selected_literal = selector.select_value(&mut context, domain_ids[0]);
         assert_eq!(
             selected_literal,
             mediator
-                .get_predicate_literal(domain_ids[0].equality_predicate(0), &assignments_integer)
+                .get_predicate_literal(domain_ids[0].equality_predicate(10), &assignments_integer)
         )
     }
 }