+++ pumpkin-lib/src/branching/value_selection/out_domain_max.rs
@@ -1,16 +1,21 @@
-use super::ValueSelector;
-use crate::basic_types::variables::IntVar;
+use crate::basic_types::DomainId;
 use crate::basic_types::Literal;
+use crate::basic_types::PredicateConstructor;
 use crate::branching::SelectionContext;
+use crate::branching::ValueSelector;
 
-/// [`ValueSelector`] which chooses to assign the provided variable to its lowest-bound.
+/// A [`ValueSelector`] which excludes the largest value from the domain.
 #[derive(Debug, Copy, Clone)]
-pub struct InDomainMin;
+pub struct OutDomainMax;
 
-impl<Var: IntVar + Copy> ValueSelector<Var> for InDomainMin {
-    fn select_value(&mut self, context: &mut SelectionContext, decision_variable: Var) -> Literal {
+impl ValueSelector<DomainId> for OutDomainMax {
+    fn select_value(
+        &mut self,
+        context: &mut SelectionContext,
+        decision_variable: DomainId,
+    ) -> Literal {
         context.get_literal_for_predicate(
-            decision_variable.upper_bound_predicate(context.lower_bound(decision_variable)),
+            decision_variable.upper_bound_predicate(context.upper_bound(decision_variable) - 1),
         )
     }
 }
@@ -19,7 +24,7 @@ impl<Var: IntVar + Copy> ValueSelector<Var> for InDomainMin {
 mod tests {
     use crate::basic_types::tests::TestRandom;
     use crate::basic_types::PredicateConstructor;
-    use crate::branching::InDomainMin;
+    use crate::branching::OutDomainMax;
     use crate::branching::SelectionContext;
     use crate::branching::ValueSelector;
 
@@ -36,13 +41,15 @@ mod tests {
         );
         let domain_ids = context.get_domains().collect::<Vec<_>>();
 
-        let mut selector = InDomainMin;
+        let mut selector = OutDomainMax;
 
         let selected_literal = selector.select_value(&mut context, domain_ids[0]);
         assert_eq!(
             selected_literal,
-            mediator
-                .get_predicate_literal(domain_ids[0].equality_predicate(0), &assignments_integer)
+            mediator.get_predicate_literal(
+                domain_ids[0].upper_bound_predicate(9),
+                &assignments_integer
+            )
         )
     }
 }