+++ pumpkin-lib/src/branching/value_selection/value_selector.rs
@@ -1,5 +1,7 @@
+use crate::basic_types::DomainId;
 use crate::basic_types::Literal;
 use crate::basic_types::PropositionalVariable;
+use crate::basic_types::Solution;
 use crate::branching::SelectionContext;
 #[cfg(doc)]
 use crate::engine::ConstraintSatisfactionSolver;
@@ -15,7 +17,7 @@ pub trait ValueSelector<Var> {
     /// otherwise should not have been selected as `decision_variable`). Returns a [`Literal`]
     /// specifying the required change in the domain (e.g. this can be retrieved using the
     /// method [`SelectionContext::get_literal_for_predicate`]).
-    fn select_value(&mut self, context: &SelectionContext, decision_variable: Var) -> Literal;
+    fn select_value(&mut self, context: &mut SelectionContext, decision_variable: Var) -> Literal;
 
     /// A function which is called after a [`Literal`] is unassigned during backtracking (i.e. when
     /// it was fixed but is no longer), specifically, it provides `literal` which is the
@@ -23,6 +25,14 @@ pub trait ValueSelector<Var> {
     /// single backtracking operation by the solver
     /// (see [`ConstraintSatisfactionSolver::backtrack`]).
     fn on_unassign_literal(&mut self, _literal: Literal) {}
+
+    /// A function which is called after a [`DomainId`] is unassigned during backtracking (i.e. when
+    /// it was fixed but is no longer), specifically, it provides `variable` which is the
+    /// [`DomainId`] which has been reset and `value` which is the value to which the variable was
+    /// previously fixed. This method could thus be called multiple times in a single
+    /// backtracking operation by the solver
+    /// (see [`backtrack`][ConstraintSatisfactionSolver::backtrack]).
+    fn on_unassign_integer(&mut self, _variable: DomainId, _value: i32) {}
 
     /// A function which is called when new [`PropositionalVariable`]s are added to the solver when
     /// encoding the objective function this method is currently only called during
@@ -33,5 +43,5 @@ pub trait ValueSelector<Var> {
     fn on_encoding_objective_function(&mut self, _all_variables: &[PropositionalVariable]) {}
 
     /// This method is called when a solution is found in the optimisation loop of [`LinearSearch`].
-    fn on_solution(&mut self, _context: &SelectionContext) {}
+    fn on_solution(&mut self, _solution: &Solution) {}
 }