+++ pumpkin-lib/src/branching/variable_selection/variable_selector.rs
@@ -1,3 +1,5 @@
+use crate::basic_types::variables::IntVar;
+use crate::basic_types::DomainId;
 use crate::basic_types::Literal;
 use crate::basic_types::PropositionalVariable;
 use crate::branching::SelectionContext;
@@ -28,9 +30,20 @@ pub trait VariableSelector<Var> {
     /// (see [`ConstraintSatisfactionSolver::backtrack`]).
     fn on_unassign_literal(&mut self, _literal: Literal) {}
 
+    /// A function which is called after a [`DomainId`] is unassigned during backtracking (i.e. when
+    /// it was fixed but is no longer), specifically, it provides `variable` which is the
+    /// [`DomainId`] which has been reset. This method could thus be called multiple times in a
+    /// single backtracking operation by the solver
+    /// (see [`backtrack`][ConstraintSatisfactionSolver::backtrack]).
+    fn on_unassign_integer(&mut self, _variable: DomainId, _value: i32) {}
+
     /// A function which is called when a [`Literal`] appears in a conflict during conflict analysis
     /// (see [`ConstraintSatisfactionSolver::compute_1uip`]).
     fn on_appearance_in_conflict_literal(&mut self, _literal: Literal) {}
+
+    /// A function which is called when a variable appears in a conflict during conflict analysis
+    /// (see [`compute_1uip`][ConstraintSatisfactionSolver::compute_1uip]).
+    fn on_appearance_in_conflict_integer(&mut self, _variable: DomainId) {}
 
     /// A function which is called when new [`PropositionalVariable`]s are added to the solver when
     /// encoding the objective function this method is currently only called during
@@ -39,7 +52,38 @@ pub trait VariableSelector<Var> {
     /// Note that this method provides **all** [`PropositionalVariable`]s and it is up to the
     /// [`VariableSelector`] to determine how to handle it.
     fn on_encoding_objective_function(&mut self, _all_variables: &[PropositionalVariable]) {}
+}
 
-    /// This method is called when a solution is found in the optimisation loop of [`LinearSearch`].
-    fn on_solution(&mut self, _context: &SelectionContext) {}
+/// Determines whether to find the [`Direction::Maximum`] or the [`Direction::Minimum`] for the
+/// [`find_extremum`] function.
+pub(crate) enum Direction {
+    Maximum,
+    Minimum,
+}
+
+/// Finds the variable with the corresponding extremum value (where the extremum is determined
+/// according to the provided [`Direction`]). The value for a variable is found using the
+/// `value_function` which takes as input a variable and returns a value.
+pub(crate) fn find_extremum<Var: IntVar + Copy, Value: Ord>(
+    variables: &[Var],
+    value_function: impl Fn(Var) -> Value,
+    context: &SelectionContext,
+    direction: Direction,
+) -> Option<Var> {
+    let filtered = variables
+        .iter()
+        .filter(|variable| !context.is_integer_fixed(**variable));
+    match direction {
+        Direction::Maximum => filtered.max_by(|x, y| {
+            let value_x = (value_function)(**x);
+            let value_y = (value_function)(**y);
+            value_x.cmp(&value_y)
+        }),
+        Direction::Minimum => filtered.min_by(|x, y| {
+            let value_x = (value_function)(**x);
+            let value_y = (value_function)(**y);
+            value_x.cmp(&value_y)
+        }),
+    }
+    .cloned()
 }